promt scrambeling v1 (only for regular factors but easily extendable for others)
this can be used to add random white spaces and random comments to the codesegment. also variable names in the code segment can be randomized. also the level and factor names can be randomized in both code and text segment (while keeping the relationship between them the same, so if a levelname in the code segment gets randomized it also gets changed to the same name in the text segment). This can be used to get a variaty of prompts out of a single template.